Extract library-file-entry primitives into modules/library_file_entries.py (PR D of quickstart.py refactor)
24 functions + 5 constants move from quickstart.py into a dedicated module:

Constants:
* LIBRARY_FILE_KINDS, LOCAL_LIBRARY_FILE_TYPES, LIBRARY_FILE_VALIDATORS,
  LIBRARY_FILE_PARSE_FUNCTIONS, SETTINGS_AUTO_SORT_HUBS_VALUES

Parsers (entry list parsing):
* _parse_library_file_entries (NEW: shared body)
* _parse_metadata_file_entries, _parse_collection_file_entries,
  _parse_overlay_file_entries (back-compat wrappers)

Slug / path utilities:
* _format_library_file_validation_error, _safe_external_artifact_slug
* _managed_library_folder_slug, _managed_library_config_root,
  _managed_library_file_root, _parse_managed_library_relative_path,
  _normalized_managed_library_relative_path, _is_bundled_library_archive_member,
  _resolve_local_library_source, _display_library_managed_location

Validation + managed-store copy:
* _validate_library_file_entry, _remove_managed_path
* _copy_library_artifact_to_managed_store, _normalize_library_external_entry
* _clone_library_file_entries_for_target, _normalize_library_file_entries_payload
* _normalize_imported_libraries_payload

List-level validators (per-kind + auto_sort_hubs):
* _validate_library_files (NEW: shared body)
* _validate_library_metadata_files, _validate_library_collection_files,
  _validate_library_overlay_files (back-compat wrappers)
* _validate_library_auto_sort_hubs

## DRY refactors landed alongside the move

The audit flagged two DRY opportunities in this cluster.  Both are
implemented here.

### 1. Three identical entry-parser functions -> one

`_parse_metadata_file_entries`, `_parse_collection_file_entries`, and
`_parse_overlay_file_entries` had byte-for-byte identical 31-line bodies
(93 lines of literal duplication).  They now share a single
`_parse_library_file_entries` implementation with thin wrappers for
backward compat with the LIBRARY_FILE_PARSE_FUNCTIONS table and external
callers.

### 2. Three near-identical list validators -> one

`_validate_library_metadata_files`, `_validate_library_collection_files`,
and `_validate_library_overlay_files` had 28-line bodies that differed
only in the kind string, the parser, the validator function, and the
payload key prefix -- all four already keyed by kind via the
LIBRARY_FILE_VALIDATORS + LIBRARY_FILE_PARSE_FUNCTIONS tables.  They
collapse into a single `_validate_library_files(libraries_data, ids,
kind)` with three one-line back-compat wrappers.

**Important monkeypatch-preservation detail**: `LIBRARY_FILE_VALIDATORS`
caches the validator function reference at import time, but tests
monkeypatch `validations.validate_metadata_file_payload` etc. on the
validations module directly.  Pre-refactor the three individual validators
called `validations.validate_metadata_file_payload(...)` by attribute
lookup on every call, so patches propagated naturally.  The shared
`_validate_library_files` preserves this behavior with
`getattr(validations, validator_attr)` on each iteration, derived from
the kind name (e.g. `metadata_files` -> `validate_metadata_file_payload`).
Three test_core_backend.py tests caught this regression on the first run
and confirm the fix is correct.

## Cross-module dependencies remaining

Three helpers stay in quickstart.py (PR G + general settings territory)
and are accessed via lazy `import quickstart as _qs` inside the routes
that need them:

* `_managed_bundle_location_for_path` -- used by `_copy_library_artifact_to_managed_store`
  (will move in PR G with the bundle-artifacts cluster)
* `_normalize_overlay_source_override_entries_payload` -- used by
  `_normalize_imported_libraries_payload` (overlay-image cluster; PR G)
* `_is_valid_auto_sort_hubs_value` -- used by `_validate_library_auto_sort_hubs`
  (tiny shared validator used by 2 non-library settings callers)

## Tests

* All 791 tests pass; 3 routes-level tests caught the monkeypatch
  regression on the first run, which fed back into the design above.
* No new test fixtures needed -- existing `qs_module.<X>` re-exports
  continue to work for tests that don't mock through the validations module.
* ruff check clean on every touched file.

## Size impact

* quickstart.py: 9,831 -> 9,313 lines (-518, -5.3%)
* modules/library_file_entries.py: 553 lines (new, under the 600 soft-limit)

Cumulative across PR A + PR C (PR #1322, open) + PR D:
11,087 -> 9,313 (-1,774, -16.0%)
or with PR C merged: -2,378 cumulative (-21.5%).
